:recycle: Move `ComponentTreeNode` in separate file
